statisticstimes 2022.05.20 v0
I accessed the data, then I sorted and organised the data using various formulae, I then analysed the data using various functionality including pivot tables. I then created a Dashboard so that the data was visualized. The file accompanies a docx which exists as a report to stakeholders wherein the visualisation is published.